pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tutto-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server implementation using FastMCP"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Leandro D Paula", email = "leandrodpaula@example.com"}
]
keywords = ["mcp", "model-context-protocol", "fastmcp", "server"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=2.13.0",
    "motor>=3.3.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "mercadopago>=2.2.0",
    "validate-docbr>=1.10.0",
    "fastapi>=0.135.3",
    "httpx>=0.27.0",
    "pyotp>=2.9.0",
    "qrcode>=7.4",
    "pillow>=10.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
tutto-mcp-server = "src.main:main"

[project.urls]
Homepage = "https://github.com/leandrodpaula/tutto-mcp-server"
Repository = "https://github.com/leandrodpaula/tutto-mcp-server"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
